empty schema data function is generic and cna live in state-machine
