+++ libs/state-machine/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/stateMachineButtonPressed';
+export * from './lib/getEmptySchemaData';
 
 // exports types from libs/state-machine/src/lib/types.ts
 // Path: libs/state-machine/src/index.ts
